requirements.txt
flask
google-cloud-storage